feat(install): install forge before 1.13

+++ src-tauri/src/install/forge/install.rs
@@ -1,55 +1,72 @@
 use std::{io::BufRead, path::PathBuf, process::Stdio};
 
+use log::{debug, info};
 use tokio::io::AsyncWriteExt;
 
-use crate::{folder::MinecraftLocation, platform::DELIMITER, DATA_LOCATION, HTTP_CLIENT};
+use crate::{platform::DELIMITER, DATA_LOCATION, HTTP_CLIENT};
 
 /// Forge Install Bootstrapper - By bangbang93
 /// [Github Repo](https://github.com/bangbang93/forge-install-bootstrapper)
 static FORGE_INSTALL_BOOTSTRAPPER: &[u8] = include_bytes!("./forge-install-bootstrapper.jar");
-
-/// Forge Install Bootstrapper - SteveXMH's Fork
-///
-/// [Github Repo](https://github.com/Steve-xmh/forge-install-bootstrapper)
-///
-/// A modified version based on [bangbang93/forge-install-bootstrapper](https://github.com/bangbang93/forge-install-bootstrapper).
-/// The purpose is to support automated installation of all versions of the installer (any version since 1.5.2 that provides an installer)
-///
-/// Usage: java -cp "forge-install-bootstrapper.jar:forge-xxx-installer.jar" com.bangbang93.ForgeInstaller "PathToDotMinecraft"
-static FORGE_INSTALL_BOOTSTRAPPER_OLD: &[u8] = include_bytes!("./forge-install-bootstrapper.jar");
 
 pub async fn install(
     install_dir: &PathBuf,
     forge_version: &str,
     mcversion: &str,
 ) -> anyhow::Result<()> {
+    let splited_forge_version: Vec<_> = forge_version.split(".").collect();
+    let bootstrapper = if splited_forge_version
+        .first()
+        .ok_or(anyhow::Error::msg("Error forge version"))?
+        .parse::<usize>()?
+        < 25
+    {
+        info!("Not using bootstrapper");
+        None
+    } else {
+        info!("Using bootstrapper");
+        Some(FORGE_INSTALL_BOOTSTRAPPER)
+    };
+    info!("Start downloading the installer");
     let installer_path = download_installer(mcversion, forge_version).await?;
+    info!("Saving bootstrapper");
     let bootstrapper_path = DATA_LOCATION
         .get()
         .unwrap()
         .temp
         .join("forge-install-bootstrapper.jar");
-    tokio::fs::write(&bootstrapper_path, FORGE_INSTALL_BOOTSTRAPPER).await?;
-    println!(
-        "\"{}:{}\"",
-        bootstrapper_path.to_str().unwrap(),
-        installer_path.to_str().unwrap()
-    );
+    if let Some(bootstrapper) = bootstrapper {
+        tokio::fs::write(&bootstrapper_path, bootstrapper).await?;
+    }
     let java = DATA_LOCATION.get().unwrap().default_jre.clone();
-    let mut command = std::process::Command::new(java)
-        .arg("-cp")
-        .arg(format!(
-            "{}{}{}",
-            bootstrapper_path.to_str().unwrap(),
-            DELIMITER,
-            installer_path.to_str().unwrap()
-        ))
-        .arg("com.bangbang93.ForgeInstaller")
-        .arg(install_dir)
-        .stdout(Stdio::piped())
-        .spawn()
-        .unwrap();
-    println!("Running forge installer");
+    info!("Starting installer");
+    let mut command =match bootstrapper {
+        Some(_) => {
+            std::process::Command::new(java)
+            .arg("-cp")
+            .arg(format!(
+                "{}{}{}",
+                bootstrapper_path.to_str().unwrap(),
+                DELIMITER,
+                installer_path.to_str().unwrap()
+            ))
+            .arg("com.bangbang93.ForgeInstaller")
+            .arg(install_dir)
+            .stdout(Stdio::piped())
+            .spawn()
+            .unwrap()
+        },
+        None => {
+            std::process::Command::new(java)
+            .arg("-jar")
+            .arg(&installer_path)
+            .arg("--installClient")
+            .arg(install_dir)
+            .stdout(Stdio::piped())
+            .spawn()
+            .unwrap()
+        }
+    };
     let out = command.stdout.take().unwrap();
     let mut out = std::io::BufReader::new(out);
     let mut buf = String::new();
@@ -60,12 +77,16 @@ pub async fn install(
         }
         if buf.ends_with("\ntrue\n") {
             success = true;
-            println!("Install success!")
+            info!("Successfully ran the fucking forge installer")
+        } else {
+            let lines: Vec<_> = buf.split("\n").collect();
+            if let Some(last) = lines.get(lines.len() - 2) {
+                debug!("{}", last);
+            }
         }
-        println!("{}", buf);
     }
     let output = command.wait_with_output().unwrap();
-    if !success || !output.status.success() {
+    if (!success && bootstrapper.is_some()) || !output.status.success() {
         tokio::fs::remove_file(installer_path).await?;
         return Err(anyhow::Error::msg("Forge installer failed"));
     }
@@ -75,6 +96,7 @@ pub async fn install(
 
 async fn download_installer(mcversion: &str, forge_version: &str) -> anyhow::Result<PathBuf> {
     let installer_url  = format!("https://maven.minecraftforge.net/net/minecraftforge/forge/{mcversion}-{forge_version}/forge-{mcversion}-{forge_version}-installer.jar");
+    info!("The installer url is: {installer_url}");
     let installer_path = DATA_LOCATION
         .get()
         .unwrap()
@@ -88,6 +110,9 @@ async fn download_installer(mcversion: &str, forge_version: &str) -> anyhow::Res
     .await?;
     let mut file = tokio::fs::File::create(&installer_path).await?;
     let response = HTTP_CLIENT.get().unwrap().get(installer_url).send().await?;
+    if !response.status().is_success() {
+        return Err(anyhow::Error::msg("Forge website return 404"));
+    }
     let src = response.bytes().await?;
     file.write_all(&src).await?;
     Ok(installer_path)
